circleci: ensure build date is consistent across build steps
CircleCI evaluates the BUILD_DATE environment variable (command)
at each and every step of the build process, so this is a hack
to ensure we use the same BUILD_DATE for metadata and tags
across both images.

The make script must run via bash to use the `source` aka `.` builtin.