default config now allows 172.18.0.0/16 for circleci
CircleCI has changed its build environment such that
containers now run with the 172.18.0.0/16 network instead
of the previous 172.17.0.0/16 network.

Before this commit, test fails because the
server (running on the new subnet) refuses the
client (also running on the new subnet) via config.

After this commit, we allow the new subnet.