make it easier to update the base image
Impact of this change
---------------------

- Renovate-bot can submit PRs to update the base image, and
  the version of freeradius gets updated automatically.

- The version scheme of the built image remains the same:
  `${VERSION}-${BUILD_DATE}-git-${HASH}`
  Users can continue to rely on explicit information in the tag.

- The image metadata (labels) is unaffected.
  Users can continue to rely on info in image labels.

Details of the change
---------------------

Before this commit, to update the base image is a 2-step process.
1. Lookup the most recent base image version and update Dockerfile.
2. Lookup the version of freeradius in the Alpine base image version
   and update `ci/build`.

After this commit, to update the base image is a 1-step process.
1. Lookup the most recent base image version and update Dockerfile.

Build installs whatever version of freeradius is in the base image
and tags the image with the version of freeradius that was installed.